Basic Scenario Parsing

+++ src/routes/game/create.ts
@@ -2,18 +2,23 @@ import express from "express";
 import formidable, {FileJSON} from 'formidable';
 import fs from "fs";
 import AdmZip from "adm-zip";
+import {unpack, UnpackingError} from "../../game/server/scenario/unpacker";
 
 const router = express.Router();
 
 router.post('/', async (req, res) => {
     const form = formidable({ multiples: true });
 
+    // Parse form data from request
     form.parse(req, async (err, fields, files) => {
+
+        // If error parsing form
         if (err) {
             res.sendStatus(400);
             return;
         }
 
+        // If no scenario file was submitted
         if (files.file == undefined) {
             res.sendStatus(400);
             return;
@@ -21,22 +26,30 @@ router.post('/', async (req, res) => {
 
         const file = files.file as unknown as FileJSON;
 
+        // Process uploaded scenario zip file
         try {
             const zip = new AdmZip(file.path);
-            const zipEntries = zip.getEntries();
-
-            zipEntries.forEach((zipEntry) => {
-                zipEntry.getDataAsync((data) => {
-                    console.log(zipEntry.entryName, data.toString());
-                })
-            });
-
-            console.log(zip.getEntry('scenario.json')?.toString());
+            let res = await unpack(zip);
+            console.log(0, res);
         }
         catch (e) {
-            res.sendStatus(400);
-            fs.unlinkSync(file.path);
-            return;
+
+            // If there was an unpacking error, return a JSON object with details
+            if (e instanceof UnpackingError) {
+                res.status(400);
+                res.send({
+                    success: false,
+                    message: e.message,
+                    context: e.context
+                });
+
+                // Cleanup uploaded zip file
+                fs.unlinkSync(file.path);
+                return;
+            }
+
+            // Otherwise rethrow error
+            throw e;
         }
 
         res.sendStatus(200);
